refactor: controller 네이밍 변경
- admin/client 기준으로 구분
- handler 위치 변경
- TODO 라인 삭제
